Add Free shipping Add threshold refactor code to make both methods separate so that it does not interfere with the payment methods interface
