+++ view/frontend/templates/pargodata.phtml
@@ -31,9 +31,9 @@
     }
 
     .pargo-btn {
-        background-color: #f0f0f0;
+        background-color: #fff200;
         color: #000;
-        border: 1px solid #f0f0f0;
+        border: 1px solid #fff200;
         padding: 8px 15px 8px 40px;
         background-image: url(https://pargo.co.za/external_img/pargo_icn_small.png);
         background-repeat: no-repeat;
@@ -44,9 +44,9 @@
     }
 
     .pargo-btn:hover {
-        background-color: #1b75bc;
+        background-color: #000;
         color: #fff;
-        border: 1px solid #1b75bc;
+        border: 1px solid #000;
         cursor: pointer;
     }
